+++ src/error.rs
@@ -97,12 +97,13 @@ pub enum Error {
 }
 
 /// An error returned from GitHub's API.
-#[derive(serde::Deserialize, Debug, Clone)]
+#[derive(Debug, Clone)]
 #[non_exhaustive]
 pub struct GitHubError {
     pub documentation_url: Option<String>,
     pub errors: Option<Vec<serde_json::Value>>,
     pub message: String,
+    pub status_code: http::StatusCode,
 }
 
 impl fmt::Display for GitHubError {