+++ tests/repos_delete_test.rs
@@ -1,30 +1,27 @@
-// Tests for calls to the /projects/{project_id} API.
 mod mock_error;
 
 use mock_error::setup_error_handler;
-use octocrab::{models::Project, Octocrab};
-use serde::{Deserialize, Serialize};
+use octocrab::Octocrab;
 use wiremock::{
     matchers::{method, path},
     Mock, MockServer, ResponseTemplate,
 };
 
-const PROJECT_ID: u32 = 1002605;
+async fn setup_delete_repo_api(template: ResponseTemplate) -> MockServer {
+    let owner: &str = "org";
+    let repo: &str = "some-repo";
 
-#[derive(Serialize, Deserialize)]
-struct FakeProject(Project);
-
-async fn setup_api(template: ResponseTemplate) -> MockServer {
     let mock_server = MockServer::start().await;
 
     Mock::given(method("DELETE"))
-        .and(path(format!("/projects/{PROJECT_ID}")))
-        .respond_with(template)
+        .and(path(format!("/repos/{owner}/{repo}")))
+        .respond_with(template.clone())
         .mount(&mock_server)
         .await;
+
     setup_error_handler(
         &mock_server,
-        &format!("DELETE on /projects/{PROJECT_ID} was not received"),
+        &format!("DELETE on /repos/{owner}/{repo} was not received"),
     )
     .await;
     mock_server
@@ -34,13 +31,17 @@ fn setup_octocrab(uri: &str) -> Octocrab {
     Octocrab::builder().base_uri(uri).unwrap().build().unwrap()
 }
 
-#[tokio::test]
-async fn should_delete_project_204() {
-    let template = ResponseTemplate::new(204);
-    let mock_server = setup_api(template).await;
+const OWNER: &str = "org";
+const REPO: &str = "some-repo";
 
+#[tokio::test]
+async fn should_delete_repo_204() {
+    let template = ResponseTemplate::new(204);
+    let mock_server = setup_delete_repo_api(template).await;
     let client = setup_octocrab(&mock_server.uri());
-    let result = client.projects().delete_project(PROJECT_ID).send().await;
+
+    let repo = client.repos(OWNER.to_owned(), REPO.to_owned());
+    let result = repo.delete().await;
 
     assert!(
         result.is_ok(),
@@ -50,12 +51,13 @@ async fn should_delete_project_204() {
 }
 
 #[tokio::test]
-async fn should_delete_project_410() {
-    let template = ResponseTemplate::new(410);
-    let mock_server = setup_api(template).await;
-
+async fn should_delete_repo_404() {
+    let template = ResponseTemplate::new(404);
+    let mock_server = setup_delete_repo_api(template).await;
     let client = setup_octocrab(&mock_server.uri());
-    let result = client.projects().delete_project(PROJECT_ID).send().await;
+
+    let repo = client.repos(OWNER.to_owned(), REPO.to_owned());
+    let result = repo.delete().await;
 
     assert!(
         result.is_err(),
@@ -65,12 +67,13 @@ async fn should_delete_project_410() {
 }
 
 #[tokio::test]
-async fn should_delete_project_500() {
+async fn should_delete_repo_500() {
     let template = ResponseTemplate::new(500);
-    let mock_server = setup_api(template).await;
-
+    let mock_server = setup_delete_repo_api(template).await;
     let client = setup_octocrab(&mock_server.uri());
-    let result = client.projects().delete_project(PROJECT_ID).send().await;
+
+    let repo = client.repos(OWNER.to_owned(), REPO.to_owned());
+    let result = repo.delete().await;
 
     assert!(
         result.is_err(),